pushing new architecture
this is a first draft of the roles/services architecture we want to
switch to. I adapted the model to this use case.

the pillar top needs to contain stuff because the extension modules are
executed after the top, so we can't iterate over node types like we can
over services in the roots top.